extract entrypoint mod
