+++ frb_codegen/src/entrypoint/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod dart;
+pub(crate) mod rust;
